requirements.txt
alembic==1.5.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
BeautifulReport==0.1.3
bleach==3.2.1
blinker==1.4
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
crypto==1.4.1
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyring==22.0.0
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==2.2.2
mistune==0.8.4
Naked==0.1.31
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numexpr==2.7.1
numpy==1.19.5
opencv-python @ file:///D:/Downloads/opencv_python-4.5.1.48-cp36-cp36m-win_amd64.whl
packaging==20.4
pandas==0.20.3
pandocfilters==1.4.3
parameterized==0.7.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qiniu==7.3.1
qtconsole==4.7.7
QtPy==1.9.0
redis==3.5.3
requests==2.25.1
requests-auth==5.2.0
scipy==1.5.4
selenium @ file:///D:/Downloads/selenium-3.6.0-py2.py3-none-any.whl
Send2Trash==1.5.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.23
svgwrite==1.4
tables==3.4.2
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
Tree==0.2.4
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
WMI==1.5.1
WTForms==2.3.3
zipp==3.4.0
